Fix eval mapping when used with unnamed clipboard
Move logic from `s:opfunc` to `scriptease#filterop` which ensures that
the unnamed clipboard is untouched. This allows the same logic to guard
against its use in the `normal! gvp` operation as well, and it
simplifies the `s:opfunc`, removing a redundant `try`/`finally`.

This fixes the `g=` mapping when used together with `set
clipboard=unnamed`.