Implement RtpAdd
A common use case for this would be editing a plugin: the user runs
  :RtpAdd
and then has their plugin directory available live for testing.

Why is this needed?

It is different from Runtime: Runtime only works for individual files.
If I have a plugin with commands that depend on autoload functions, or a
colorscheme, Runtime doesn't suit me. In the first case, I have to call
Runtime twice. In the second, I may not want to immediately run the
colorscheme (or it may *need* to be in the runtimepath). A third case:
the user wants to verify their plugin's documentation is correct.
  :RtpAdd | help my-plugin
should now load the live version (not the version installed in ~/.vim,
for example).

An additional benefit is that Ve and friends now automatically learn
about the plugin files in my current directory (the plugin I'm working
on).